Rename the command to `eksctl experimental export-templates`
