+++ pkg/ctl/utils/utils.go
@@ -16,7 +16,6 @@ func Command(flagGrouping *cmdutils.FlagGrouping) *cobra.Command {
 	cmdutils.AddResourceCmd(flagGrouping, verbCmd, updateKubeProxyCmd)
 	cmdutils.AddResourceCmd(flagGrouping, verbCmd, updateAWSNodeCmd)
 	cmdutils.AddResourceCmd(flagGrouping, verbCmd, updateCoreDNSCmd)
-	cmdutils.AddResourceCmd(flagGrouping, verbCmd, exportCmd)
 
 	return verbCmd
 }
